requirements.txt
python-dotenv
pyyaml
langchain
langchain-openai
faiss-cpu
streamlit

